+++ packages/evokit/__tests__/createBlock.test.js
@@ -45,6 +45,16 @@ describe("import { createBlock } from 'evokit';", () => {
             expect(shallow(<Block block-modName='modValue' />).contains(<div className='ek-block ek-block_modName_modValue'></div>)).toBeTruthy();
         });
 
+        it('with className', () => {
+            expect(
+                shallow(
+                    <Block block-modName='modValue' className='class-name' />
+                ).contains(
+                    <div className='ek-block ek-block_modName_modValue class-name'></div>
+                )
+            ).toBeTruthy();
+        });
+
         describe("custom class prefix", () => {
             it('default', () => {
                 expect(shallow(<Block block-preset={{ b: 'ek-' }} />).contains(<div className='ek-block'></div>)).toBeTruthy();
@@ -53,6 +63,58 @@ describe("import { createBlock } from 'evokit';", () => {
             it('mods', () => {
                 expect(shallow(<Block block-preset={{ b: 'ek-' }} block-modName='modValue' />)
                     .contains(<div className='ek-block ek-block_modName_modValue'></div>)).toBeTruthy();
+            });
+        });
+
+        describe("css modules", () => {
+            it('default', () => {
+                expect(
+                    shallow(
+                        <Block block-preset={{ b: '', css: { block: 'h67f' } }} />
+                    ).contains(
+                        <div className='h67f'></div>
+                    )
+                ).toBeTruthy();
+            });
+
+            it('with mods', () => {
+                expect(
+                    shallow(
+                        <Block
+                            block-preset={{ b: '', css: { block: 'h67f', block_modName_modValue: 'ko45' } }}
+                            block-modName='modValue'
+                        />
+                    ).contains(
+                        <div className='h67f ko45'></div>
+                    )
+                ).toBeTruthy();
+            });
+
+            it('with className', () => {
+                expect(
+                    shallow(
+                        <Block
+                            block-preset={{ b: '', css: { block: 'h67f' } }}
+                            className='class-name'
+                        />
+                    ).contains(
+                        <div className='h67f class-name'></div>
+                    )
+                ).toBeTruthy();
+            });
+
+            it('with mods and className', () => {
+                expect(
+                    shallow(
+                        <Block
+                            block-preset={{ b: '', css: { block: 'h67f', block_modName_modValue: 'ko45' } }}
+                            block-modName='modValue'
+                            className='class-name VasYa'
+                        />
+                    ).contains(
+                        <div className='h67f ko45 class-name VasYa'></div>
+                    )
+                ).toBeTruthy();
             });
         });
     });